P77: Cherry-pick from P77-camera-dist-precision-fix to P77-Copter-4.2.2

+++ libraries/AP_Camera/AP_Camera.h
@@ -46,7 +46,8 @@ public:
     void            control(float session, float zoom_pos, float zoom_step, float focus_lock, float shooting_cmd, float cmd_id);
 
     // set camera trigger distance in a mission
-    void            set_trigger_distance(uint32_t distance_m)
+    // P77: change uint32_t -> float
+    void            set_trigger_distance(float distance_m)
     {
         _trigg_dist.set(distance_m);
     }
